src/accounts/pyproject.toml
[tool.poetry]
name = "accounts"
version = "0.1.0"
description = "Сервис accounts проекта vidos. Отвечает за работу с пользовательскими данными. Также является сервисом авторизации."
authors = ["Andreev Aisen  <isenandreev@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "3.2.23"
djangorestframework = "^3.14.0"
markdown = "^3.5.1"
django-filter = "^23.5"
python-dotenv = "^1.0.0"
gunicorn = "^21.2.0"
drf-yasg = "^1.21.7"
djoser = "^2.2.2"
django-cors-headers = "^4.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
